+++ src/main/java/roomescape/reservation/Reservation.java
@@ -7,6 +7,10 @@ public class Reservation {
     private String date;
     private String time;
 
+    public Reservation() {
+
+    }
+
     public Reservation(long id, String name, String date, String time) {
         this.id = id;
         this.name = name;
@@ -28,6 +32,10 @@ public class Reservation {
 
     public String getTime() {
         return time;
+    }
+
+    public void setId(long id) {
+        this.id = id;
     }
 
     public void setName(String name) {
