+++ src/main/java/roomescape/reservation/ReservationDTOMapper.java
@@ -4,11 +4,21 @@ package roomescape.reservation;
 public class ReservationDTOMapper {
 
     public static Reservation toEntity(ReservationDTO reservationDTO) {
-        return new Reservation(reservationDTO.getId(), reservationDTO.getName(), reservationDTO.getDate(), reservationDTO.getTime());
+        Reservation reservation = new Reservation();
+        reservation.setId(reservationDTO.getId());
+        reservation.setName(reservationDTO.getName());
+        reservation.setDate(reservationDTO.getDate());
+        reservation.setTime(reservationDTO.getTime());
+        return reservation;
     }
 
     public static ReservationDTO toDTO(Reservation reservation) {
-        return new ReservationDTO(reservation.getId(), reservation.getName(), reservation.getDate(), reservation.getTime());
+        ReservationDTO reservationDTO = new ReservationDTO();
+        reservationDTO.setId(reservation.getId());
+        reservationDTO.setName(reservation.getName());
+        reservationDTO.setDate(reservation.getDate());
+        reservationDTO.setTime(reservation.getTime());
+        return reservationDTO;
     }
 
 }
