+++ src/main/java/roomescape/reservation/ReservationExceptionHandler.java
@@ -1,0 +1,20 @@
+package roomescape.reservation;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import roomescape.reservation.exception.NotFoundReservationException;
+
+@ControllerAdvice
+public class ReservationExceptionHandler {
+
+    @ExceptionHandler(NotFoundReservationException.class)
+    public ResponseEntity<String> handleException(NotFoundReservationException e) {
+        return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().build();
+    }
+}
