+++ src/main/java/roomescape/reservation/exception/NotFoundReservationException.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.exception;
 
 public class NotFoundReservationException extends RuntimeException {
